Rename `database.plugin` to `db.plugin`
