+++ src/plugins/auth.plugin.ts
@@ -5,7 +5,7 @@ import Elysia, { NotFoundError, t } from 'elysia'
 import { ENV } from '~/config'
 import { ApiErrorSchema } from '~/utils/response.util'
 import { AuthenticationError } from '~/utils/errors.util'
-import { db } from './database.plugin'
+import { db } from './db.plugin'
 import { logger } from './logger.plugin'
 
 export const auth = betterAuth({
