+++ src/plugins/db.plugin.ts
@@ -10,7 +10,7 @@ export const db = drizzle({
   connection: {
     host: ENV.DB_HOST,
     port: ENV.DB_PORT,
-    user: ENV.DB_USER,
+    username: ENV.DB_USER,
     password: ENV.DB_PASS,
     database: ENV.DB_NAME,
     ssl:
