+++ test/setup.ts
@@ -1,13 +1,14 @@
 import { afterAll, beforeAll } from 'bun:test'
 import type { DrizzleError } from 'drizzle-orm'
-import { migrator } from '~/plugins/database.plugin'
+import { migrator } from '~/plugins/db.plugin'
 
 beforeAll(async () => {
   try {
     await migrator()
   } catch (error) {
     const err = error as DrizzleError
-    const errCode = (err.cause as { code: string } | undefined)?.code ?? 'UNKNOWN'
+    const errCode =
+      (err.cause as { code: string } | undefined)?.code ?? 'UNKNOWN'
     console.error('Error during database migration:', errCode)
   }
 })
